move WEBHOOK_SECRET_KEY loading and validation to webhook app
SOFTWARE-3325
